+++ pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.metamx</groupId>
     <artifactId>http-client</artifactId>
-    <version>0.9.13-SNAPSHOT</version>
+    <version>1.0.0-SNAPSHOT</version>
     <name>http-client</name>
     <description>An HTTP client using Netty</description>
     <scm>
@@ -109,6 +109,14 @@
             </plugin>
         </plugins>
     </build>
+
+    <repositories>
+        <repository>
+            <id>pub-libs</id>
+            <name>pub-libs-local</name>
+            <url>https://metamx.artifactoryonline.com/metamx/pub-libs-releases-local</url>
+        </repository>
+    </repositories>
 
     <distributionManagement>
         <repository>
