+++ src/main/java/com/metamx/http/client/CredentialedHttpClient.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2015 Yahoo!
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.metamx.http.client;
+
+import com.google.common.util.concurrent.ListenableFuture;
+import com.metamx.http.client.auth.Credentials;
+import com.metamx.http.client.response.HttpResponseHandler;
+import org.jboss.netty.handler.codec.http.HttpMethod;
+
+import java.net.URL;
+
+/**
+ */
+public class CredentialedHttpClient implements HttpClient
+{
+
+  private final Credentials creds;
+  private final HttpClient delegate;
+
+  public CredentialedHttpClient(Credentials creds, HttpClient delegate)
+  {
+    this.creds = creds;
+    this.delegate = delegate;
+  }
+
+  @Override
+  public <Intermediate, Final> ListenableFuture<Final> go(Request request, HttpResponseHandler<Intermediate, Final> handler) {
+    return delegate.go(creds.addCredentials(request), handler);
+  }
+}
